fix[db]: deprecation warning, absolute paths and typing import
Added deprecation warning for ``s2.db`` in favor of ``s2.store``; db (for databse) was misleading, as this module does not allow queries in the way a db would. Instead, the functionality of the module is moved to ``s2.store`` where classes will be referred to as DataStores which is less misleading. ``s2.store`` will also address the issues in ``s2.db`` of providing a single interface for both ``S2Paper`` and ``S2Author`` while playing nice with mypy.